fix: change 04-copy-directory

+++ 04-copy-directory/index.js
@@ -3,18 +3,20 @@ const path = require('path');
 const currentDir = path.join(__dirname, 'files');
 const copyDir = path.join(__dirname, 'files-copy');
 
-fs.access(copyDir, fs.constants.F_OK, (error) => {
-  console.log(`${copyDir} ${error ? 'not exist' : 'exist'}`);
+fs.rm(copyDir, { recursive: true }, (error) => {
+  if (error) return console.error(error.message);
   fs.mkdir(copyDir, { recursive: true }, (error) => {
     if (error) return console.error(error.message);
     fs.readdir(currentDir, { withFileTypes: true }, (error, result) => {
       if (error) return console.error(error.message);
       result.forEach(file => {
-        fs.copyFile(path.join(currentDir, file.name), path.join(copyDir, file.name), (error) => {
-          if (error) return console.error(error.message);
-          console.log('File copied successfully');
-        });
+        if (file.isFile()) {
+          fs.copyFile(path.join(currentDir, file.name), path.join(copyDir, file.name), (error) => {
+            if (error) return console.error(error.message);
+          });
+        }
       });
+      console.log('File copied successfully');
     });
   });
 });
